+++ src/websockets/server.py
@@ -5,7 +5,7 @@ import binascii
 import email.utils
 import http
 import warnings
-from typing import Any, Generator, List, Optional, Sequence, Tuple, cast
+from typing import Any, Callable, Generator, List, Optional, Sequence, Tuple, cast
 
 from .datastructures import Headers, MultipleValuesError
 from .exceptions import (
@@ -58,6 +58,10 @@ class ServerProtocol(Protocol):
             should be tried.
         subprotocols: list of supported subprotocols, in order of decreasing
             preference.
+        select_subprotocol: callback for selecting a subprotocol among
+            those supported by the client and the server. It has the same
+            signature as the :meth:`select_subprotocol` method, including a
+            :class:`ServerProtocol` instance as first argument.
         state: initial state of the WebSocket connection.
         max_size: maximum size of incoming messages in bytes;
             :obj:`None` to disable the limit.
@@ -73,6 +77,7 @@ class ServerProtocol(Protocol):
         origins: Optional[Sequence[Optional[Origin]]] = None,
         extensions: Optional[Sequence[ServerExtensionFactory]] = None,
         subprotocols: Optional[Sequence[Subprotocol]] = None,
+        select_subprotocol: Optional[SelectSubprotocol] = None,
         state: State = CONNECTING,
         max_size: Optional[int] = 2**20,
         logger: Optional[LoggerLike] = None,
@@ -86,6 +91,14 @@ class ServerProtocol(Protocol):
         self.origins = origins
         self.available_extensions = extensions
         self.available_subprotocols = subprotocols
+        if select_subprotocol is not None:
+            # Bind select_subprotocol then shadow self.select_subprotocol.
+            # Use setattr to work around https://github.com/python/mypy/issues/2427.
+            setattr(
+                self,
+                "select_subprotocol",
+                select_subprotocol.__get__(self, self.__class__),
+            )
 
     def accept(self, request: Request) -> Response:
         """
@@ -96,7 +109,7 @@ class ServerProtocol(Protocol):
 
         You must send the handshake response with :meth:`send_response`.
 
-        You can modify it before sending it, for example to add HTTP headers.
+        You may modify it before sending it, for example to add HTTP headers.
 
         Args:
             request: WebSocket handshake request event received from the client.
@@ -175,7 +188,8 @@ class ServerProtocol(Protocol):
         return Response(101, "Switching Protocols", headers)
 
     def process_request(
-        self, request: Request
+        self,
+        request: Request,
     ) -> Tuple[str, Optional[str], Optional[str]]:
         """
         Check a handshake request and negotiate extensions and subprotocol.
@@ -273,6 +287,7 @@ class ServerProtocol(Protocol):
            Optional[Origin]: origin, if it is acceptable.
 
         Raises:
+            InvalidHandshake: if the Origin header is invalid.
             InvalidOrigin: if the origin isn't acceptable.
 
         """
@@ -323,7 +338,7 @@ class ServerProtocol(Protocol):
             HTTP response header and list of accepted extensions.
 
         Raises:
-            InvalidHandshake: to abort the handshake with an HTTP 400 error.
+            InvalidHandshake: if the Sec-WebSocket-Extensions header is invalid.
 
         """
         response_header_value: Optional[str] = None
@@ -383,60 +398,79 @@ class ServerProtocol(Protocol):
            also the value of the ``Sec-WebSocket-Protocol`` response header.
 
         Raises:
-            InvalidHandshake: to abort the handshake with an HTTP 400 error.
+            InvalidHandshake: if the Sec-WebSocket-Subprotocol header is invalid.
 
         """
-        subprotocol: Optional[Subprotocol] = None
+        subprotocols: Sequence[Subprotocol] = sum(
+            [
+                parse_subprotocol(header_value)
+                for header_value in headers.get_all("Sec-WebSocket-Protocol")
+            ],
+            [],
+        )
 
-        header_values = headers.get_all("Sec-WebSocket-Protocol")
-
-        if header_values and self.available_subprotocols:
-
-            parsed_header_values: List[Subprotocol] = sum(
-                [parse_subprotocol(header_value) for header_value in header_values], []
-            )
-
-            subprotocol = self.select_subprotocol(
-                parsed_header_values, self.available_subprotocols
-            )
-
-        return subprotocol
+        return self.select_subprotocol(subprotocols)
 
     def select_subprotocol(
         self,
-        client_subprotocols: Sequence[Subprotocol],
-        server_subprotocols: Sequence[Subprotocol],
+        subprotocols: Sequence[Subprotocol],
     ) -> Optional[Subprotocol]:
         """
         Pick a subprotocol among those offered by the client.
 
-        If several subprotocols are supported by the client and the server,
-        the default implementation selects the preferred subprotocols by
-        giving equal value to the priorities of the client and the server.
+        If several subprotocols are supported by both the client and the server,
+        pick the first one in the list declared the server.
 
-        If no common subprotocol is supported by the client and the server, it
-        proceeds without a subprotocol.
+        If the server doesn't support any subprotocols, continue without a
+        subprotocol, regardless of what the client offers.
 
-        This is unlikely to be the most useful implementation in practice, as
-        many servers providing a subprotocol will require that the client uses
-        that subprotocol.
+        If the server supports at least one subprotocol and the client doesn't
+        offer any, abort the handshake with an HTTP 400 error.
+
+        You provide a ``select_subprotocol`` argument to :class:`ServerProtocol`
+        to override this logic. For example, you could accept the connection
+        even if client doesn't offer a subprotocol, rather than reject it.
+
+        Here's how to negotiate the ``chat`` subprotocol if the client supports
+        it and continue without a subprotocol otherwise::
+
+            def select_subprotocol(protocol, subprotocols):
+                if "chat" in subprotocols:
+                    return "chat"
 
         Args:
-            client_subprotocols: list of subprotocols offered by the client.
-            server_subprotocols: list of subprotocols available on the server.
+            subprotocols: list of subprotocols offered by the client.
 
         Returns:
-            Optional[Subprotocol]: Subprotocol, if a common subprotocol was
-            found.
+            Optional[Subprotocol]: Selected subprotocol, if a common subprotocol
+            was found.
+
+            :obj:`None` to continue without a subprotocol.
+
+        Raises:
+            NegotiationError: custom implementations may raise this exception
+                to abort the handshake with an HTTP 400 error.
 
         """
-        subprotocols = set(client_subprotocols) & set(server_subprotocols)
-        if not subprotocols:
+        # Server doesn't offer any subprotocols.
+        if not self.available_subprotocols:  # None or empty list
             return None
-        priority = lambda p: (
-            client_subprotocols.index(p) + server_subprotocols.index(p)
+
+        # Server offers at least one subprotocol but client doesn't offer any.
+        if not subprotocols:
+            raise NegotiationError("missing subprotocol")
+
+        # Server and client both offer subprotocols. Look for a shared one.
+        proposed_subprotocols = set(subprotocols)
+        for subprotocol in self.available_subprotocols:
+            if subprotocol in proposed_subprotocols:
+                return subprotocol
+
+        # No common subprotocol was found.
+        raise NegotiationError(
+            "invalid subprotocol; expected one of "
+            + ", ".join(self.available_subprotocols)
         )
-        return sorted(subprotocols, key=priority)[0]
 
     def reject(
         self,
@@ -517,6 +551,12 @@ class ServerProtocol(Protocol):
             self.events.append(request)
 
         yield from super().parse()
+
+
+SelectSubprotocol = Callable[
+    [ServerProtocol, Sequence[Subprotocol]],
+    Optional[Subprotocol],
+]
 
 
 class ServerConnection(ServerProtocol):